+++ src/PrismicRichText/Serialize.svelte
@@ -1,12 +1,17 @@
 <script lang="ts">
-	import type { asTree } from "@prismicio/richtext";
+	import type { asTree } from "@prismicio/client/richtext";
 
-	import type { SvelteRichTextSerializer } from "../types";
+	import type { RichTextComponent, RichTextComponents } from "../types";
 
 	import DefaultComponent from "./DefaultComponent.svelte";
+	import Serialize from "./Serialize.svelte";
 
-	export let components: SvelteRichTextSerializer = {};
-	export let children: ReturnType<typeof asTree>["children"];
+	type Props = {
+		components: RichTextComponents;
+		children: ReturnType<typeof asTree>["children"];
+	};
+
+	const { components, children }: Props = $props();
 
 	const CHILD_TYPE_RENAMES = {
 		"list-item": "listItem",
@@ -17,20 +22,21 @@
 
 	function getComponent(child: ReturnType<typeof asTree>["children"][number]) {
 		return (
-			components[
+			(components[
 				CHILD_TYPE_RENAMES[child.type as keyof typeof CHILD_TYPE_RENAMES] ||
 					(child.type as keyof typeof components)
-			] || DefaultComponent
+			] as RichTextComponent) || DefaultComponent
 		);
 	}
 </script>
 
-{#each children as child}
-	<svelte:component this={getComponent(child)} node={child.node}>
+{#each children as child (child.key)}
+	{@const Component = getComponent(child)}
+	<Component node={child.node}>
 		<!-- This formatting is intentional to prevent unwanted whitespace between elements. -->
-		{#if child.children.length > 0}<svelte:self
+		{#if child.children.length > 0}<Serialize
 				children={child.children}
 				{components}
-			/>{/if}</svelte:component
+			/>{/if}</Component
 	>
 {/each}