+++ test/PrismicLink.test.ts
@@ -13,8 +13,8 @@ it("renders the link's text if no children are provided", (ctx) => {
 	const field = ctx.mock.value.link({ type: "Web", model, withText: true });
 	const { container } = render(PrismicLinkTestWrapper, { field });
 
-	expect(container.innerHTML).toBe(
-		`<div><a href="${field.url}" rel="noreferrer">${field.text}</a></div>`,
+	expect(container.innerHTML.replaceAll("<!---->", "")).toBe(
+		`<a href="${field.url}" rel="noreferrer">${field.text}</a>`,
 	);
 });
 
@@ -27,7 +27,7 @@ it("renders the given children, overriding the link's text", (ctx) => {
 	const children = ctx.mock.value.keyText();
 	const { container } = render(PrismicLinkTestWrapper, { field, children });
 
-	expect(container.innerHTML).toBe(
-		`<div><a href="${field.url}" rel="noreferrer">${children}</a></div>`,
+	expect(container.innerHTML.replaceAll("<!---->", "")).toBe(
+		`<a href="${field.url}" rel="noreferrer">${children}</a>`,
 	);
 });