+++ test/PrismicLinkTestWrapper.svelte
@@ -5,6 +5,7 @@ passing slot content to the PrismicLink component is not easy in tests.
 
 <script lang="ts">
 	import { FilledLinkToWebField } from "@prismicio/client";
+
 	import { PrismicLink } from "../src";
 
 	export let field: FilledLinkToWebField;