+++ vitest.config.ts
@@ -11,4 +11,10 @@ export default defineConfig({
 		},
 		setupFiles: ["./test/__setup__.ts"],
 	},
+	// Tell Vitest to use the `browser` entry points in `package.json` files, even though it's running in Node
+	resolve: process.env.VITEST
+		? {
+				conditions: ["browser"],
+			}
+		: undefined,
 });